+++ src/components/Header/components/HeaderSearchButton.test.tsx
@@ -9,12 +9,16 @@ import * as envUtils from "../../../utils";
 
 jest.mock("../../../utils", () => ({
   getEnvVar: jest.fn(),
+  sendAnalyticsNavClickEvent: jest.fn(),
 }));
 
 describe("HeaderSearchButton Accessibility", () => {
   beforeAll(() => {
     (envUtils.getEnvVar as jest.Mock).mockImplementation((key) =>
-      key === "VITE_APP_ENV" ? "qa" : ""
+      key === "VITE_APP_ENV" ? "qa" : "",
+    );
+    (envUtils.sendAnalyticsNavClickEvent as jest.Mock).mockImplementation(
+      () => {},
     );
   });
   it("passes axe accessibility test", async () => {
@@ -31,7 +35,7 @@ describe("HeaderSearchButton Accessibility", () => {
 describe("HeaderSearchButton", () => {
   beforeAll(() => {
     (envUtils.getEnvVar as jest.Mock).mockImplementation((key) =>
-      key === "VITE_APP_ENV" ? "qa" : ""
+      key === "VITE_APP_ENV" ? "qa" : "",
     );
   });
   describe("Desktop", () => {